Вынос класса опций в отдельный файл
